requirements.txt
feedparser
requests
beautifulsoup4
lxml
python-dotenv
schedule